Implement #25 - Expand/hide all button
This commit adds functionality to expand and hide
all entries in the table.
This is achieved by adding a button in the header
of the table, which triggers the toggleAll() function.
A helper function is added to check if all elements
are visible, so the icon of the button can be changed
accordingly.